refactor: move typeguards into util
